Grid System - Implement Moving Between Rooms
NOTE: There is a bug currently in the system where
the new grid copies the state of Units left alive
in the old grid leading to an occupied tile with
no unit on it in the new grid. This should get
fixed automatically once we implement code to
drop additional Player units when leaving a room.